Refactor application directory structure
Seperated modules out for readability's sake.

Namely, moved each route into it's own file, cut down on global imports,
sorted imports/dependenies, and split out tests.

NW
