+++ metrix/src/db.rs
@@ -1,0 +1,10 @@
+use std::env;
+
+use diesel::prelude::*;
+
+pub fn establish_connection() -> PgConnection {
+    let database_url = env::var("DATABASE_URL")
+        .expect("DATABASE_URL must be set");
+    PgConnection::establish(&database_url)
+        .expect(&format!("Error connecting to {}", database_url))
+}
