+++ metrix/src/main.rs
@@ -1,26 +1,10 @@
-#![feature(proc_macro_hygiene, decl_macro)]
-#[macro_use] extern crate rocket;
-#[macro_use] extern crate diesel;
-#[macro_use] extern crate serde;
 #[macro_use] extern crate diesel_migrations;
-#[macro_use] extern crate nom;
-extern crate chrono;
-extern crate rocket_contrib;
-extern crate serde_json;
-
 embed_migrations!();
 
-pub mod lib;
-pub mod models;
-pub mod schema;
-pub mod app;
-pub mod parser;
-
-use lib::establish_connection;
-
-pub fn create_app() -> rocket::Rocket {
-  app::create_app()
-}
+use metrix::{
+	db::establish_connection,
+	create_app,
+};
 
 fn main() {
     println!("### Enter the Metrix ###");
@@ -32,5 +16,3 @@ fn main() {
 
     create_app().launch();
 }
-
-#[cfg(test)] mod tests;
