+++ metrix/src/routes/mod.rs
@@ -1,0 +1,2 @@
+pub mod ping;
+pub mod metrics;
