+++ metrix/src/routes/ping.rs
@@ -1,0 +1,4 @@
+#[get("/")]
+pub fn ping() -> &'static str {
+    "pong"
+}
