+++ metrix/tests/parser.rs
@@ -1,19 +1,6 @@
-use super::create_app;
-use rocket::local::Client;
-use rocket::http::Status;
+use metrix::parser::parse_query_string;
+use metrix::parser::parse_parameter_name;
 
-use crate::parser::parse_query_string;
-use crate::parser::parse_parameter_name;
-
-#[test]
-fn ping_me_baby() {
-  let client = Client::new(create_app()).expect("valid rocket instance");
-
-  let mut response = client.get("/ping").dispatch();
-
-  assert_eq!(response.status(), Status::Ok);
-  assert_eq!(response.body_string(), Some("pong".into()));
-}
 
 #[test]
 fn parse_query_string_single_expression_integer_value() {
